Proyecto inventario-11

+++ src/main/java/cervantes/inventario/controlador/ProductoControlador.java
@@ -63,6 +63,10 @@ public class ProductoControlador {
             @PathVariable int id,
             @RequestBody Producto productoRecibido){
         Producto producto = this.productoServicio.buscarProductoPorId(id);
+
+        if (producto==null)
+            throw new RecursoNoEncontradoExcepcion("No se encontro el id: "+ id);
+
         producto.setDescripcion(productoRecibido.getDescripcion());
         producto.setPrecio(productoRecibido.getPrecio());
         producto.setExistencia(productoRecibido.getExistencia());
@@ -74,6 +78,9 @@ public class ProductoControlador {
     @DeleteMapping("/productos/{id}")
     public ResponseEntity<Map<String, Boolean>> eliminarProducto(@PathVariable int id){
         Producto producto = productoServicio.buscarProductoPorId(id);
+        if(producto ==null)
+            throw  new RecursoNoEncontradoExcepcion("No se encontro el id: " + id);
+
         this.productoServicio.eliminarProductoPorId(producto.getIdProducto());
         Map<String,Boolean> respuesta = new HashMap<>();
         respuesta.put("eliminado", Boolean.TRUE);
